Refactor lazy instrumentation loader (#1060)
* Refactor lazy instrumentation loader

* Remove unnecessary interface implementation
